backend/financial_data_collector/requirements.txt
claude-agent-sdk>=0.1.45
yfinance>=1.2.0
requests>=2.32.0
python-dotenv>=1.2.0
